refactor: noda_platform.h -> noda/platform.h
