+++ noda-library/src/noda/platform.h
